reorganize code for packaging
