+++ .gitignore
@@ -1,0 +1,43 @@
+*.py[co]
+*~
+# Packages
+*.egg
+*.egg-info
+dist
+build
+eggs
+parts
+var
+sdist
+develop-eggs
+.installed.cfg
+.eggs
+
+# Installer logs
+pip-log.txt
+
+# Unit test / coverage reports
+.coverage
+.tox
+
+#Translations
+*.mo
+
+#Mr Developer
+.mr.developer.cfg
+
+#Vim
+*.swp
+
+#Code coverage
+cover
+
+#eclipse/pydev
+.project
+.pydevproject
+
+# compiled docs
+docs/_build
+
+# Python virtualenv
+.venv
