+++ README.md
@@ -1,0 +1,7 @@
+# Tripal API Library
+
+A Python script to generate html reports for [Stacks](http://catchenlab.life.illinois.edu/stacks/) results
+
+## License
+
+Available under the GPL License
